verification to avoid wrong insert to all methods

+++ controllers/tasklistCRUD.ts
@@ -3,15 +3,25 @@ import { db } from "../src/db/database";
 import { tasklist } from "../src/db/schema";
 import type { Request, response, Response } from "express";
 
-async function checkIfExists(id) {
+async function taskExists(id: number) {
   const [checkValue] = await db
     .select()
     .from(tasklist)
     .where(eq(tasklist.id, id));
-  console.log(checkValue);
+  return checkValue;
 }
 
-export const getTask = async (req, res) => {
+function isValidId(id: number) {
+  let parsedId = Number(id);
+  return Number.isInteger(parsedId) && parsedId > 0;
+}
+
+async function findTask(id: number) {
+  return await db.select().from(tasklist).where(eq(tasklist.id, id));
+}
+
+///////////////*******CRUD OF MY TASKS++++++++++ */
+export const getTask = async (req: Request, res: Response) => {
   try {
     const tasks = await db.select().from(tasklist);
     res.json(tasks);
@@ -20,66 +30,97 @@ export const getTask = async (req, res) => {
   }
 };
 
-export const getTaskById = async (req, res) => {
+export const getTaskById = async (req: Request, res: Response) => {
   try {
     const id = Number(req.params.id);
-
-    if (Number.isInteger(id) && id > 0) {
-      checkIfExists(id);
-      const singleTask = await db
-        .select()
-        .from(tasklist)
-        .where(eq(tasklist.id, id));
-      res.json(singleTask);
-      res.end();
-    } else {
+    if (!isValidId(id)) {
       res.status(400).json({
-        id: "The parametrer id inserted is not a number or was not found in the database",
+        error: "The given id is not valid",
       });
     }
+    let taskExist = await taskExists(id);
+    if (!taskExist) {
+      res
+        .status(404)
+        .json({ error: "The task you're looking for doesn't exist" });
+    }
+    const singleTask = await findTask(id);
+    res.json(singleTask);
   } catch (error) {
     console.log(error);
   }
 };
 
-export const postTask = async (req, res) => {
+export const addTask = async (req: Request, res: Response) => {
   try {
     const { title, description } = await req.body;
-    const [task] = await db.insert(tasklist).values({ title, description });
-    res.status(201).json(task);
+
+    if (
+      typeof title === "string" &&
+      title.length > 0 &&
+      title.trim() &&
+      typeof description === "string" &&
+      description.length > 0 &&
+      description.trim()
+    ) {
+      let cleanTitle = title.trim();
+      let cleanDescription = description.trim();
+      const [result] = await db
+        .insert(tasklist)
+        .values({ title: cleanTitle, description: cleanDescription });
+      const singleTask = await findTask(result.insertId);
+
+      res.status(201).json(singleTask);
+    } else {
+      res.status(400).json({
+        error: "The given titles and descriptions of your task are not valid",
+      });
+    }
   } catch (error) {
     console.log(error);
   }
 };
 
 /**==================================UPDATE */
-export const updateTask = async (req, res) => {
+export const updateTask = async (req: Request, res: Response) => {
   try {
     const id = Number(req.params.id);
     const { title, description } = req.body;
 
-    if (Number.isInteger(id) && id > 0) {
+    if (!isValidId(id)) {
+      res.status(400).json({
+        id: "The given Id is not valid",
+      });
+    }
+    let returnedTask = await taskExists(id);
+
+    if (returnedTask && title.trim() && description.trim()) {
+      let cleanTitle = title.trim();
+      let cleanDescription = description.trim();
       const [task] = await db
         .update(tasklist)
-        .set({ title, description })
+        .set({ title: cleanTitle, description: cleanDescription })
         .where(eq(tasklist.id, id));
-      res.status(200).json(task);
+      let [taskfound] = await findTask(id);
+      console.log({ task, taskfound });
+      res.status(200).json(taskfound);
     } else {
-      res.status(400).json({
-        id: "The parametrer id inserted is not a number or was not found in the database",
-      });
+      res.status(404).json({ error: "Nothing was found" });
     }
   } catch (error) {
     console.log(error);
   }
 };
+
 ///////////////////*************Delete  */
 export const deleteTask = async (req: Request, res: Response) => {
   try {
     let id = Number(req.params.id);
-    if (Number.isInteger(id) && id > 0) {
+    let confirmTaskExists = await taskExists(id);
+
+    if (Number.isInteger(id) && id > 0 && confirmTaskExists) {
       await db.delete(tasklist).where(eq(tasklist.id, id));
-      res.status(204).end();
+      res.status(204).json({ success: "Task deleted successfully" }).end();
     } else {
       res.status(400).json({
         id: "The parametrer id inserted is not a number or was not found in the database",
